fixes issue #227 (WIP)
